+++ api/grpcserver/activation_service_test.go
@@ -17,6 +17,7 @@ import (
 	"github.com/spacemeshos/go-spacemesh/common/types"
 	"github.com/spacemeshos/go-spacemesh/events"
 	"github.com/spacemeshos/go-spacemesh/sql"
+	"github.com/spacemeshos/go-spacemesh/sql/statesql"
 )
 
 func Test_Highest_ReturnsGoldenAtxOnError(t *testing.T) {
@@ -156,7 +157,7 @@ func TestGet_IdentityCanceled(t *testing.T) {
 	atxProvider := grpcserver.NewMockatxProvider(ctrl)
 	activationService := grpcserver.NewActivationService(atxProvider, types.ATXID{1})
 
-	smesher, proof := grpcserver.BallotMalfeasance(t, sql.InMemory())
+	smesher, proof := grpcserver.BallotMalfeasance(t, statesql.InMemory())
 	previous := types.RandomATXID()
 	id := types.RandomATXID()
 	atx := types.ActivationTx{