+++ atxsdata/warmup_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/spacemeshos/go-spacemesh/sql/atxs"
 	"github.com/spacemeshos/go-spacemesh/sql/layers"
 	"github.com/spacemeshos/go-spacemesh/sql/mocks"
+	"github.com/spacemeshos/go-spacemesh/sql/statesql"
 )
 
 func gatx(
@@ -38,7 +39,7 @@ func gatx(
 func TestWarmup(t *testing.T) {
 	types.SetLayersPerEpoch(3)
 	t.Run("sanity", func(t *testing.T) {
-		db := sql.InMemory()
+		db := statesql.InMemory()
 		applied := types.LayerID(10)
 		nonce := types.VRFPostIndex(1)
 		data := []types.ActivationTx{
@@ -61,19 +62,19 @@ func TestWarmup(t *testing.T) {
 		}
 	})
 	t.Run("no data", func(t *testing.T) {
-		c, err := Warm(sql.InMemory(), 1, zaptest.NewLogger(t))
+		c, err := Warm(statesql.InMemory(), 1, zaptest.NewLogger(t))
 		require.NoError(t, err)
 		require.NotNil(t, c)
 	})
 	t.Run("closed db", func(t *testing.T) {
-		db := sql.InMemory()
+		db := statesql.InMemory()
 		require.NoError(t, db.Close())
 		c, err := Warm(db, 1, zaptest.NewLogger(t))
 		require.Error(t, err)
 		require.Nil(t, c)
 	})
 	t.Run("db failures", func(t *testing.T) {
-		db := sql.InMemory()
+		db := statesql.InMemory()
 		nonce := types.VRFPostIndex(1)
 		data := gatx(types.ATXID{1, 1}, 1, types.NodeID{1}, nonce)
 		require.NoError(t, atxs.Add(db, &data, types.AtxBlob{}))