+++ checkpoint/recovery_collecting_deps_test.go
@@ -9,15 +9,15 @@ import (
 
 	"github.com/spacemeshos/go-spacemesh/activation/wire"
 	"github.com/spacemeshos/go-spacemesh/common/types"
-	"github.com/spacemeshos/go-spacemesh/sql"
 	"github.com/spacemeshos/go-spacemesh/sql/atxs"
 	"github.com/spacemeshos/go-spacemesh/sql/identities"
+	"github.com/spacemeshos/go-spacemesh/sql/statesql"
 )
 
 func TestCollectingDeps(t *testing.T) {
 	golden := types.RandomATXID()
 	t.Run("collect marriage ATXs", func(t *testing.T) {
-		db := sql.InMemory()
+		db := statesql.InMemory()
 
 		marriageATX := &wire.ActivationTxV1{
 			InnerActivationTxV1: wire.InnerActivationTxV1{