+++ tortoise/tortoise_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/spacemeshos/go-spacemesh/sql/builder"
 	"github.com/spacemeshos/go-spacemesh/sql/certificates"
 	"github.com/spacemeshos/go-spacemesh/sql/layers"
+	"github.com/spacemeshos/go-spacemesh/sql/statesql"
 	"github.com/spacemeshos/go-spacemesh/tortoise/opinionhash"
 	"github.com/spacemeshos/go-spacemesh/tortoise/sim"
 )
@@ -336,7 +337,7 @@ func tortoiseFromSimState(tb testing.TB, state sim.State, opts ...Opt) *recovery
 	return &recoveryAdapter{
 		TB:       tb,
 		Tortoise: trtl,
-		db:       state.DB.Executor,
+		db:       state.DB.Database,
 		atxdata:  state.Atxdata,
 	}
 }
@@ -467,7 +468,7 @@ func TestComputeExpectedWeight(t *testing.T) {
 	} {
 		t.Run(tc.desc, func(t *testing.T) {
 			var (
-				db     = sql.InMemory()
+				db     = statesql.InMemory()
 				epochs = map[types.EpochID]*epochInfo{}
 				first  = tc.target.Add(1).GetEpoch()
 			)